Diseño principal de la estructura de dos jugadores

+++ client_client.cpp
@@ -5,7 +5,7 @@
 #include "server_TaTeTi.h"
 #include "common_socket.h"
 
-// #include "Action.cpp"
+#include "common_Accion.h"
 
 int main(int argc, char *argv[]) {
     Socket socket;
@@ -23,15 +23,18 @@ int main(int argc, char *argv[]) {
             std::cin >> name;
             socket.send(Unirse(name));;
         } else if (!actionStr.compare("jugar")) {
-            uint8_t i, j;
+            int i, j;
             std::cin >> i >> j;
-            socket.send(Jugar(i,j));
+            socket.send(Jugar(i, j));
         }
         // Imprimir respuesta. Si es el caso de crear, 
         // va a tener que esperar de todas formas, asi que sale.
     }
     return 0;
-
+    // uint8_t hexa = 0x12;
+    // int x = (hexa >> 4) & 0x0F;
+    // int y = hexa & 0x0F;
+    // std::cerr << "hexa<" << hexa << "> x<" << x << "> y<" << y << ">" << std::endl;
     // uint8_t res = 0, a = 1, b = 2;
     // res = b | (a << 4);
     // uint8_t expected = 0x12;
